+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New cvar `mdd_projection`. Determines how angles will be projected on the screen. This affects both CampingGaz-HUD and Snap-HUD.
 
 ## [1.2.0] - 2020-04-13
 ### Added